% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_function.R
\name{find_medians}
\alias{find_medians}
\title{Calculate Medians for Numeric Columns}
\usage{
find_medians(data)
}
\arguments{
\item{data}{A data frame with numeric columns to calculate medians.}
}
\value{
A named vector of medians for each numeric column.
}
\description{
This function computes the median for each numeric column in a dataset, ignoring NA values.
}
