% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_function.R
\name{replace_missing_with_median}
\alias{replace_missing_with_median}
\title{Replace Missing Values with Medians}
\usage{
replace_missing_with_median(data)
}
\arguments{
\item{data}{A data frame with numeric columns to clean by replacing missing values.}
}
\value{
A data frame with missing values replaced by the median of each column.
}
\description{
This function replaces missing values in each numeric column with the column’s median.
}
